@@ -91,6 +91,6 @@ public class ServiceAreaCalculatorDefinition extends GeoEventProcessorDefinition
 	@Override
 	public String getVersion()
 	{
-		return "10.5.0";
+		return "10.6.0";
 	}
 }
